Update
Changed some font sizes, fixed the bug that you can't start a new game after the end, and adjusted some bits and pieces. Problems that can also be adjusted: no level numbering, new meteorites will appear on the ship causing a direct loss of a life. The color of the high score interface is not uniform. There is time to summarize the code of controller.java to reduce redundancy.